+++ src/read.c
@@ -5289,10 +5289,9 @@ avifResult avifDecoderParse(avifDecoder * decoder)
 {
     avifDiagnosticsClearError(&decoder->diag);
 
-    // Color only or alpha only is not currently supported.
-    if ((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) != 0 &&
-        (decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) != AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) {
-        avifDiagnosticsPrintf(&decoder->diag, "imageContentToDecode set to only color or only alpha is not supported");
+    // Alpha only is not currently supported.
+    if ((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) == AVIF_IMAGE_CONTENT_ALPHA) {
+        avifDiagnosticsPrintf(&decoder->diag, "imageContentToDecode set to only alpha is not supported");
         return AVIF_RESULT_NOT_IMPLEMENTED;
     }
     if (!decoder->io || !decoder->io->read) {
@@ -6095,10 +6094,9 @@ avifResult avifDecoderReset(avifDecoder * decoder)
 
     memset(&decoder->ioStats, 0, sizeof(decoder->ioStats));
 
-    // Color only or alpha only is not currently supported.
-    if ((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) != 0 &&
-        (decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) != AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) {
-        avifDiagnosticsPrintf(&decoder->diag, "imageContentToDecode set to only color or only alpha is not supported");
+    // Alpha only is not currently supported.
+    if ((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) == AVIF_IMAGE_CONTENT_ALPHA) {
+        avifDiagnosticsPrintf(&decoder->diag, "imageContentToDecode set to only alpha is not supported");
         return AVIF_RESULT_NOT_IMPLEMENTED;
     }
 
@@ -6295,19 +6293,21 @@ avifResult avifDecoderReset(avifDecoder * decoder)
         colorCodecType = codecType[AVIF_ITEM_COLOR];
 
         // Optional alpha auxiliary item
-        avifBool isAlphaItemInInput;
-        AVIF_CHECKRES(avifMetaFindAlphaItem(data->meta,
-                                            mainItems[AVIF_ITEM_COLOR],
-                                            &data->tileInfos[AVIF_ITEM_COLOR],
-                                            &mainItems[AVIF_ITEM_ALPHA],
-                                            &data->tileInfos[AVIF_ITEM_ALPHA],
-                                            &isAlphaItemInInput));
-        if (mainItems[AVIF_ITEM_ALPHA]) {
-            AVIF_CHECKRES(avifDecoderItemReadAndParse(decoder,
-                                                      mainItems[AVIF_ITEM_ALPHA],
-                                                      isAlphaItemInInput,
-                                                      &data->tileInfos[AVIF_ITEM_ALPHA].grid,
-                                                      &codecType[AVIF_ITEM_ALPHA]));
+        avifBool isAlphaItemInInput = AVIF_FALSE;
+        if (decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_ALPHA) {
+            AVIF_CHECKRES(avifMetaFindAlphaItem(data->meta,
+                                                mainItems[AVIF_ITEM_COLOR],
+                                                &data->tileInfos[AVIF_ITEM_COLOR],
+                                                &mainItems[AVIF_ITEM_ALPHA],
+                                                &data->tileInfos[AVIF_ITEM_ALPHA],
+                                                &isAlphaItemInInput));
+            if (mainItems[AVIF_ITEM_ALPHA]) {
+                AVIF_CHECKRES(avifDecoderItemReadAndParse(decoder,
+                                                          mainItems[AVIF_ITEM_ALPHA],
+                                                          isAlphaItemInInput,
+                                                          &data->tileInfos[AVIF_ITEM_ALPHA].grid,
+                                                          &codecType[AVIF_ITEM_ALPHA]));
+            }
         }
 
         // Section 10.2.6 of 23008-12:2024/AMD 1:2024(E):
@@ -6332,7 +6332,7 @@ avifResult avifDecoderReset(avifDecoder * decoder)
 
         // AVIF_ITEM_SAMPLE_TRANSFORM (not used through mainItems because not a coded item (well grids are not coded items either but it's different)).
         avifDecoderItem * const sampleTransformItem = avifDecoderDataFindSampleTransformImageItem(data);
-        if ((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) &&
+        if ((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR) &&
             (decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_SAMPLE_TRANSFORMS) && sampleTransformItem != NULL) {
             AVIF_ASSERT_OR_RETURN(data->sampleTransformNumInputImageItems == 0);
 
@@ -6397,6 +6397,9 @@ avifResult avifDecoderReset(avifDecoder * decoder)
                                                           &codecType[*category]));
 
                 // Optional alpha auxiliary item
+                if (!(decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_ALPHA)) {
+                    continue;
+                }
                 avifBool isAlphaInputImageItemInInput = AVIF_FALSE;
                 AVIF_CHECKRES(avifMetaFindAlphaItem(data->meta,
                                                     mainItems[*category],
@@ -6463,13 +6466,19 @@ avifResult avifDecoderReset(avifDecoder * decoder)
 
             AVIF_CHECKRES(avifDecoderAdoptGridTileCodecTypeIfNeeded(decoder, mainItems[c], &data->tileInfos[c]));
 
-            if (c == AVIF_ITEM_COLOR || c == AVIF_ITEM_ALPHA) {
-                if (!(decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA)) {
+            if (c == AVIF_ITEM_COLOR) {
+                if (!(decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR)) {
                     continue;
                 }
-            } else if (c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_0_COLOR || c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_1_COLOR ||
-                       c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_0_ALPHA || c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_1_ALPHA) {
-                AVIF_ASSERT_OR_RETURN((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA) &&
+            } else if (c == AVIF_ITEM_ALPHA) {
+                if (!(decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_ALPHA)) {
+                    continue;
+                }
+            } else if (c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_0_COLOR || c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_1_COLOR) {
+                AVIF_ASSERT_OR_RETURN((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR) &&
+                                      (decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_SAMPLE_TRANSFORMS));
+            } else if (c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_0_ALPHA || c == AVIF_ITEM_SAMPLE_TRANSFORM_INPUT_1_ALPHA) {
+                AVIF_ASSERT_OR_RETURN((decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_ALPHA) &&
                                       (decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_SAMPLE_TRANSFORMS));
             } else {
                 AVIF_ASSERT_OR_RETURN(c == AVIF_ITEM_GAIN_MAP);
@@ -7038,7 +7047,7 @@ avifResult avifDecoderNextImage(avifDecoder * decoder)
     }
 
     // If decoder->data->tileInfos[AVIF_ITEM_COLOR].tileCount == 0, it means
-    // decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA was equal to 0.
+    // decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR was equal to 0.
     // Only apply Sample Transforms if there is a color item to apply it onto.
     if (decoder->data->tileInfos[AVIF_ITEM_COLOR].tileCount != 0 && decoder->data->meta->sampleTransformExpression.count > 0) {
         AVIF_CHECKRES(avifDecoderApplySampleTransform(decoder, decoder->image));
@@ -7213,7 +7222,7 @@ static uint32_t avifGetDecodedRowCount(const avifDecoder * decoder, const avifTi
 uint32_t avifDecoderDecodedRowCount(const avifDecoder * decoder)
 {
     if (decoder->data->tileInfos[AVIF_ITEM_COLOR].tileCount == 0) {
-        // decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR_AND_ALPHA
+        // decoder->imageContentToDecode & AVIF_IMAGE_CONTENT_COLOR
         // was likely 0 when avifDecoderNextImage() was called.
         // avifDecoderDecodedRowCount() only describes decoder->image->yuvPlanes[0].
         // There is no available luma plane, so return 0 decoded rows.